New functionality
In order to standardize the mapping in impairmentCheck() across all servos, a calibration() function was added, which changed two new VarSpeedServo parameters according to how the servo moves. Also implemented feedbackPin value into the impairmentCheck() and wait() functions, to remove the need to pass the feedback pin value in as a parameter.