Diminishing average function
The new function uses a diminishing average, where previous actualMovement values affect the current value by a factor of (1/2)^(n+1), where n = number of loops since the loop where that value was created.  With a reduction in actualMovementThreshold generated from experimental data, the movement is now almost false positive free, and the false positives only sometimes occur on the first check after a new movement has just begun (so this is likely due to inertia).