mapping function and calibration function changes
In an attempt to refine the feedback data, I changed the mapping function from the STL version to a custom one to reduce the integer rounding problem. I also slightly increased the difference threshold. These two changes have allowed me to successfully run the program with limited testing and no false positives in two of the servos.